rename BoardPage to DatabasePage
