--- components/[pageId]/BoardPage/index.ts
@@ -1,3 +1,0 @@
-import BoardPage from './BoardPage';
-
-export default BoardPage;
