+++ components/[pageId]/DatabasePage/DatabasePage.tsx
