+++ components/[pageId]/DatabasePage/index.ts
@@ -1,0 +1,3 @@
+import BoardPage from './DatabasePage';
+
+export default BoardPage;
